Correct widescreen aspect ratio
The program was modifying the output height as if the original aspect ratio was 1:1. This is incorrect, and show take into account the current aspect ratio of 4:3 before modifying.

Explaining the maths behind this: You should start with the target aspect ratio and divide it by the current aspect ratio to determine what factor the adjustment needs. This works out to be 3 by 4 in this case. What's interesting about the 16:9 aspect ratio is that it's actually 4²:3², so the proper tweaking ratio makes sense there too. If you're working with pixels, a 640×480 4:3 resolution becomes 640×360 in 16:9. You can divide the width and height by their aspect ratio to verify the numbers are all right.